refactor(timeline): use native sticky instead of scroll-mirroring for header strip
Removing overflow-y-hidden from the scroll container lets position: sticky
work relative to the page scroll directly, so TimeScale no longer needs
a translateX(-scrollLeft) mirror or a scrollLeft prop - it goes back to
tracking its own scroll container internally, same as on main.
